# src/api/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Authy",
    "version": "1.0.0",
    "description": "An express api for authentication using JWTs.",
    "license": {
      "name": "MIT",
      "url": "https://spdx.org/licenses/MIT.html"
    },
    "contact": {
      "name": "Authy Kishor",
      "email": "kishorliv@gmail.com"
    },
    "servers": [
      {
        "url": "http://localhost:3000",
        "description": "Development server"
      }
    ]
  },
  "paths": {
    "/accounts/register": {
      "post": {
        "summary": "Register a new account and send a verification email",
        "description": "If the user is already registered and not verified, email verification is sent again.",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Doe"
                  },
                  "email": {
                    "type": "string",
                    "example": "john@doe.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "foo@foo"
                  },
                  "confirmPassword": {
                    "type": "string",
                    "example": "foo@foo"
                  }
                },
                "required": [
                  "firstName",
                  "lastName",
                  "email",
                  "password",
                  "confirmPassword"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful account registration and email has been sent for verification.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Registration successful. Please check your email for verification."
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}